Add bigSave: 完成

+++ gmall-pms/src/main/java/com/atguigu/gmall/pms/vo/SkuVo.java
@@ -60,7 +60,7 @@ public class SkuVo extends SkuEntity {
     /**
      * 满几件
      */
-    private Integer fullCount;
+    private BigDecimal fullCount;
     /**
      * 打几折
      */
